fix: dried up the action test code
