+++ test/actions/mintNFT.test.ts
@@ -1,24 +1,19 @@
 import { Keypair } from '@solana/web3.js';
-import axios, { AxiosResponse } from 'axios';
-import { Connection } from '../../src';
-import { NodeWallet, Wallet } from '../../src';
+import axios from 'axios';
+import { Connection, NodeWallet, Wallet } from '../../src';
 import { mintNFT } from '../../src/actions';
 import { FEE_PAYER } from '../utils';
 import { MasterEdition, Metadata } from '../../src/programs/metadata';
+import { mockAxios200, mockAxios404, uri } from './shared';
 
 jest.mock('axios');
 
-const mockedAxiosGet = axios.get as jest.MockedFunction<typeof axios>;
-const uri = 'https://bafkreibj4hjlhf3ehpugvfy6bzhhu2c7frvyhrykjqmoocsvdw24omfqga.ipfs.dweb.link';
-
 describe('minting an NFT', () => {
-  let connection: Connection;
+  const connection = new Connection('devnet');
+  const wallet = new NodeWallet(FEE_PAYER);
   let mint: Keypair;
-  let wallet: Wallet;
 
   beforeAll(() => {
-    connection = new Connection('devnet');
-    wallet = new NodeWallet(FEE_PAYER);
     jest
       .spyOn(connection, 'sendRawTransaction')
       .mockResolvedValue(
@@ -33,32 +28,7 @@ describe('minting an NFT', () => {
 
   describe('when can find metadata json', () => {
     beforeEach(() => {
-      const mockedResponse: AxiosResponse = {
-        data: {
-          name: 'Holo Design (0)',
-          symbol: '',
-          description:
-            'A holo of some design in a lovely purple, pink, and yellow. Pulled from the Internet. Demo only.',
-          seller_fee_basis_points: 100,
-          image:
-            'https://bafybeidq34cu23fq4u57xu3hp2usqqs7miszscyu4kjqyjo3hv7xea6upe.ipfs.dweb.link',
-          external_url: '',
-          properties: {
-            creators: [
-              {
-                address: wallet.publicKey.toString(),
-                share: 100,
-              },
-            ],
-          },
-        },
-        status: 200,
-        statusText: 'OK',
-        headers: {},
-        config: {},
-      };
-
-      mockedAxiosGet.mockResolvedValue(mockedResponse);
+      mockAxios200(wallet);
     });
 
     test('generates a unique mint and creates metadata plus master edition from metadata URL and max supply', async () => {
@@ -82,15 +52,7 @@ describe('minting an NFT', () => {
 
   describe('when metadata json not found', () => {
     beforeEach(() => {
-      const mockedResponse: AxiosResponse = {
-        data: {},
-        status: 404,
-        statusText: 'NOT FOUND',
-        headers: {},
-        config: {},
-      };
-
-      mockedAxiosGet.mockRejectedValue(mockedResponse);
+      mockAxios404();
     });
 
     test('exits the action and throws an error', async () => {
