+++ test/actions/updateMetadata.test.ts
@@ -1,53 +1,23 @@
 import { Keypair } from '@solana/web3.js';
-import axios, { AxiosResponse } from 'axios';
-import { Account, Connection, NodeWallet, Wallet } from '../../src';
+import axios from 'axios';
+import { Account, Connection, NodeWallet } from '../../src';
 import { mintNFT } from '../../src/actions';
 import { FEE_PAYER, pause } from '../utils';
 import { Creator, Metadata, MetadataDataData } from '../../src/programs/metadata';
 import { updateMetadata } from '../../src/actions/updateMetadata';
+import { mockAxios200, uri } from './shared';
 
 jest.mock('axios');
 jest.setTimeout(100000);
 
-const mockedAxiosGet = axios.get as jest.MockedFunction<typeof axios>;
-const uri = 'https://bafkreibj4hjlhf3ehpugvfy6bzhhu2c7frvyhrykjqmoocsvdw24omfqga.ipfs.dweb.link';
-
 describe('updating metadata on a master edition', () => {
-  let connection: Connection;
+  const connection = new Connection('devnet');
+  const wallet = new NodeWallet(FEE_PAYER);
   let mint: Keypair;
-  let wallet: Wallet;
-
-  beforeAll(() => {
-    connection = new Connection('devnet');
-    wallet = new NodeWallet(FEE_PAYER);
-  });
 
   beforeEach(() => {
     mint = Keypair.generate();
-    const mockedResponse: AxiosResponse = {
-      data: {
-        name: 'Holo Design (0)',
-        symbol: '',
-        description:
-          'A holo of some design in a lovely purple, pink, and yellow. Pulled from the Internet. Demo only.',
-        seller_fee_basis_points: 100,
-        image: 'https://bafybeidq34cu23fq4u57xu3hp2usqqs7miszscyu4kjqyjo3hv7xea6upe.ipfs.dweb.link',
-        external_url: '',
-        properties: {
-          creators: [
-            {
-              address: wallet.publicKey.toString(),
-              share: 100,
-            },
-          ],
-        },
-      },
-      status: 200,
-      statusText: 'OK',
-      headers: {},
-      config: {},
-    };
-    mockedAxiosGet.mockResolvedValue(mockedResponse);
+    mockAxios200(wallet);
   });
 
   test('updates successfully', async () => {
